+++ lib/src/cli/agent_cli.dart
@@ -61,6 +61,8 @@ import '../model_roles/model_roles.dart';
 import '../model_roles/vision_models.dart';
 import '../providers/chatgpt_oauth.dart';
 import '../providers/codemie_sso.dart';
+import '../providers/copilot_device_flow.dart';
+import '../providers/copilot_oauth.dart';
 import '../providers/dial.dart';
 import '../providers/models_endpoint.dart';
 import '../providers/openrouter_oauth.dart';